Add support for post-finalize calls to cnmemFree.
In C++, there may still be live objects when we reach time to call
cnmemFinalize. Thus, we need a mecanism to ensure correctness. This commit
adds a way to deal with that situation. Each object which may want to
call cnmemFree *after* cnmemFinalize has to increase the reference counter
on the library (say in its constructor) with cnmemRetain and release the
counter when it is sure it won't call cnmemFree anymore (say in its
destructor). If all the objects respect that rule, it should work as
expected.

This commit also removes the CNMEM_STATUS_MEMORY_LEAK status since it is
now impossible to report such an error with the described mecanism.