Fix the warning caused by the visibility attribute
According to GNU documentation, __attribute__(visibility ("default"))
has to be set before a type. In the case of the cnmemGetErrorString, we
have to set that attribute before the star of const char * to avoid the
warning. Thus, it should be

const char CNMEM_API * cnmemGetErrorString(...)

and not

const char* CNMEM_API cnmemGetErrorString(...)

This commit fixes that issue.